backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.7.1
django-cors-headers==3.13.0
autopep8==1.7.0
django-dotenv==1.4.2
PyJWT==2.6.0
django-filter==2.4.0
django-redis==4.12.1
Pillow==9.5.0